+++ Editor/PluginWindow.h
@@ -69,7 +69,7 @@ namespace ToolKit
 
       void AddResolutionName(const String& name);
       void RemoveResolutionName(const String& name);
-      void RemoveResolutionName(int index);
+      void RemoveResolutionName(size_t index);
 
       int m_numDefaultResNames  = 0;
